+++ assets/_src/modules/full-offload.tsx
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { MouseEvent, useContext } from 'react';
+import { MouseEvent, useContext, useState } from 'react';
 import { mdiImageMultipleOutline } from '@mdi/js';
 
 /**
@@ -17,7 +17,14 @@ import Card from '../components/card';
 import ProgressBar from '../components/progress';
 
 const FullOffload = () => {
+	const [action, setAction] = useState('');
 	const { inProgress, setInProgress } = useContext(SettingsContext);
+
+	const runAction = (e: MouseEvent, actionName: string) => {
+		e.preventDefault();
+		setAction(actionName);
+		setInProgress(true);
+	};
 
 	return (
 		<Card
@@ -39,15 +46,32 @@ const FullOffload = () => {
 					)}
 				</p>
 
-				{inProgress && <ProgressBar action="full-remove" />}
+				{inProgress && <ProgressBar action={action} />}
 
-				<button
-					className="button is-small"
-					onClick={(e) => setInProgress(true)}
-					disabled={inProgress}
-				>
-					{__('Bulk delete', 'cf-images')}
-				</button>
+				<div className="buttons mb-1">
+					<button
+						className="button is-small"
+						onClick={(e) => runAction(e, 'full-remove')}
+						disabled={inProgress}
+					>
+						{__('Bulk delete', 'cf-images')}
+					</button>
+
+					<button
+						className="button is-small"
+						onClick={(e) => runAction(e, 'full-restore')}
+						disabled={inProgress}
+					>
+						{__('Bulk restore', 'cf-images')}
+					</button>
+				</div>
+
+				<p>
+					{__(
+						'* Bulk actions will remove/restore physical files from the WordPress media library. Images that have not been offloaded to Cloudflare will be skipped.',
+						'cf-images'
+					)}
+				</p>
 			</div>
 		</Card>
 	);